requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Azure OpenAI
openai==1.12.0
tiktoken==0.5.2

# Vector Store
qdrant-client==1.7.3

# Document Processing
pypdf==4.0.1
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5

# Text Processing
sentence-transformers==2.3.1
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
aiofiles==23.2.1

# Logging
loguru==0.7.2